gatewayapicontroller: Don't use BeforeAll/AfterAll
Replace BeforeAll and AfterAll with BeforeEach and AfterEach,
respectively, and add explicit logic to prevent the cleanup logic
in AfterEach from running before the tests are all done.

Because of the way that Origin tests are run, BeforeAll and AfterAll
actually behave as BeforeEach and AfterEach.  That is, an AfterAll block
is run not once after all the specs in the Describe container are done
but rather once after each spec, and similarly, a BeforeAll block is run
before each spec.  Before this change, the cleanup logic was happening
after every spec, meaning OSSM was being uninstalled and re-installed
between each test, which caused a lot of churn and flakiness.

This change replaces BeforeAll and AfterAll because using them is
misleading.  This change also makes the tests faster and more reliable.

* gatewayapicontroller.go (gatewayClassName): Move declaration from the
Describe block to the package level so that helpers can use it.
(ossmAndOLMResourcesCreated, defaultGatewayclassAccepted)
(customGatewayclassAccepted, lbAndServiceAndDnsrecordAreCreated)
(httprouteObjectCreated, gieEnabled): New consts used to identify tests.
(testNames): New variable with a list of all tests in the file.
(Describe): Replace BeforeAll and AfterAll with BeforeEach and
AfterEach.  Use checkAllTestsDone in the AfterEach block to skip cleanup
while tests are still running.  Use a deferred call to markTestDone in
each test's It block to mark the test as done when it finishes.
(annotationKeyForTest): New helper.  Return the key for an annotation on
the default gatewayclass that indicates whether a given test is done.
(markTestDone): New helper.  Add an annotation to the default
gatewayclass to indicate that a test is done.
(checkAllTestsDone): New helper. Check the default gatewayclass's
annotations and return a Boolean value indicating whether all tests are
done.

+++ test/extended/router/gatewayapicontroller.go
@@ -30,13 +30,6 @@ import (
 	gatewayapiv1 "sigs.k8s.io/gateway-api/apis/v1"
 )
 
-var (
-	requiredCapabilities = []configv1.ClusterVersionCapability{
-		configv1.ClusterVersionCapabilityMarketplace,
-		configv1.ClusterVersionCapabilityOperatorLifecycleManager,
-	}
-)
-
 const (
 	// Max time duration for the DNS resolution
 	dnsResolutionTimeout = 10 * time.Minute
@@ -47,6 +40,36 @@ const (
 	ingressNamespace = "openshift-ingress"
 	// istioName is the name of the Istio CR.
 	istioName = "openshift-gateway"
+	// The name of the default gatewayclass, which is used to install OSSM.
+	gatewayClassName = "openshift-default"
+
+	ossmAndOLMResourcesCreated         = "ensure-resources-are-created"
+	defaultGatewayclassAccepted        = "ensure-default-gatewayclass-is-accepted"
+	customGatewayclassAccepted         = "ensure-custom-gatewayclass-is-accepted"
+	lbAndServiceAndDnsrecordAreCreated = "ensure-lb-and-service-and-dnsrecord-are-created"
+	httprouteObjectCreated             = "ensure-httproute-object-is-created"
+	gieEnabled                         = "ensure-gie-is-enabled"
+)
+
+var (
+	requiredCapabilities = []configv1.ClusterVersionCapability{
+		configv1.ClusterVersionCapabilityMarketplace,
+		configv1.ClusterVersionCapabilityOperatorLifecycleManager,
+	}
+	// testNames is a list of names that are used to track when tests are
+	// done in order to check whether it is safe to clean up resources that
+	// these tests share, such as the gatewayclass and Istio CR.  These
+	// names are embedded within annotation keys of the form test-%s-done.
+	// Because annotation keys are limited to 63 characters, each of these
+	// names must be no longer than 53 characters.
+	testNames = []string{
+		ossmAndOLMResourcesCreated,
+		defaultGatewayclassAccepted,
+		customGatewayclassAccepted,
+		lbAndServiceAndDnsrecordAreCreated,
+		httprouteObjectCreated,
+		gieEnabled,
+	}
 )
 
 var _ = g.Describe("[sig-network-edge][OCPFeatureGate:GatewayAPIController][Feature:Router][apigroup:gateway.networking.k8s.io]", g.Ordered, g.Serial, func() {
@@ -68,14 +91,12 @@ var _ = g.Describe("[sig-network-edge][OCPFeatureGate:GatewayAPIController][Feat
 		expectedSubscriptionSource = "redhat-operators"
 		// The expected OSSM operator namespace.
 		expectedSubscriptionNamespace = "openshift-operators"
-		// The gatewayclass name used to create ossm and other gateway api resources.
-		gatewayClassName = "openshift-default"
 		// gatewayClassControllerName is the name that must be used to create a supported gatewayClass.
 		gatewayClassControllerName = "openshift.io/gateway-controller/v1"
 		//OSSM Deployment Pod Name
 		deploymentOSSMName = "servicemesh-operator3"
 	)
-	g.BeforeAll(func() {
+	g.BeforeEach(func() {
 		isokd, err := isOKD(oc)
 		if err != nil {
 			e2e.Failf("Failed to get clusterversion to determine if release is OKD: %v", err)
@@ -100,56 +121,62 @@ var _ = g.Describe("[sig-network-edge][OCPFeatureGate:GatewayAPIController][Feat
 		}
 	})
 
-	g.AfterAll(func() {
-		g.By("Deleting the gateways")
+	g.AfterEach(func() {
+		if !checkAllTestsDone(oc) {
+			e2e.Logf("Skipping cleanup while not all GatewayAPIController tests are done")
+		} else {
+			g.By("Deleting the gateways")
 
-		for _, name := range gateways {
-			err = oc.AdminGatewayApiClient().GatewayV1().Gateways(ingressNamespace).Delete(context.Background(), name, metav1.DeleteOptions{})
-			o.Expect(err).NotTo(o.HaveOccurred(), "Gateway %s could not be deleted", name)
-		}
+			for _, name := range gateways {
+				err = oc.AdminGatewayApiClient().GatewayV1().Gateways(ingressNamespace).Delete(context.Background(), name, metav1.DeleteOptions{})
+				o.Expect(err).NotTo(o.HaveOccurred(), "Gateway %s could not be deleted", name)
+			}
 
-		g.By("Deleting the GatewayClass")
+			g.By("Deleting the GatewayClass")
 
-		if err := oc.AdminGatewayApiClient().GatewayV1().GatewayClasses().Delete(context.Background(), gatewayClassName, metav1.DeleteOptions{}); err != nil && !apierrors.IsNotFound(err) {
-			e2e.Failf("Failed to delete GatewayClass %q", gatewayClassName)
-		}
+			if err := oc.AdminGatewayApiClient().GatewayV1().GatewayClasses().Delete(context.Background(), gatewayClassName, metav1.DeleteOptions{}); err != nil && !apierrors.IsNotFound(err) {
+				e2e.Failf("Failed to delete GatewayClass %q", gatewayClassName)
+			}
 
-		g.By("Deleting the Istio CR")
+			g.By("Deleting the Istio CR")
 
-		// Explicitly deleting the Istio CR should not strictly be
-		// necessary; the Istio CR has an owner reference on the
-		// gatewayclass, and so deleting the gatewayclass should cause
-		// the garbage collector to delete the Istio CR.  However, it
-		// has been observed that the Istio CR sometimes does not get
-		// deleted, and so we have an explicit delete command here just
-		// in case.  The --ignore-not-found option should prevent errors
-		// if garbage collection has already deleted the object.
-		o.Expect(oc.AsAdmin().WithoutNamespace().Run("delete").Args("--ignore-not-found=true", "istio", istioName).Execute()).Should(o.Succeed())
+			// Explicitly deleting the Istio CR should not strictly be
+			// necessary; the Istio CR has an owner reference on the
+			// gatewayclass, and so deleting the gatewayclass should cause
+			// the garbage collector to delete the Istio CR.  However, it
+			// has been observed that the Istio CR sometimes does not get
+			// deleted, and so we have an explicit delete command here just
+			// in case.  The --ignore-not-found option should prevent errors
+			// if garbage collection has already deleted the object.
+			o.Expect(oc.AsAdmin().WithoutNamespace().Run("delete").Args("--ignore-not-found=true", "istio", istioName).Execute()).Should(o.Succeed())
 
-		g.By("Waiting for the istiod pod to be deleted")
+			g.By("Waiting for the istiod pod to be deleted")
 
-		o.Eventually(func(g o.Gomega) {
-			podsList, err := oc.AdminKubeClient().CoreV1().Pods(ingressNamespace).List(context.Background(), metav1.ListOptions{LabelSelector: "app=istiod"})
-			g.Expect(err).NotTo(o.HaveOccurred())
-			g.Expect(podsList.Items).Should(o.BeEmpty())
-		}).WithTimeout(10 * time.Minute).WithPolling(10 * time.Second).Should(o.Succeed())
+			o.Eventually(func(g o.Gomega) {
+				podsList, err := oc.AdminKubeClient().CoreV1().Pods(ingressNamespace).List(context.Background(), metav1.ListOptions{LabelSelector: "app=istiod"})
+				g.Expect(err).NotTo(o.HaveOccurred())
+				g.Expect(podsList.Items).Should(o.BeEmpty())
+			}).WithTimeout(10 * time.Minute).WithPolling(10 * time.Second).Should(o.Succeed())
 
-		g.By("Deleting the OSSM Operator resources")
+			g.By("Deleting the OSSM Operator resources")
 
-		operator, err := operatorsv1.NewForConfigOrDie(oc.AsAdmin().UserConfig()).Operators().Get(context.Background(), serviceMeshOperatorName, metav1.GetOptions{})
-		o.Expect(err).NotTo(o.HaveOccurred(), "Failed to get Operator %q", serviceMeshOperatorName)
+			operator, err := operatorsv1.NewForConfigOrDie(oc.AsAdmin().UserConfig()).Operators().Get(context.Background(), serviceMeshOperatorName, metav1.GetOptions{})
+			o.Expect(err).NotTo(o.HaveOccurred(), "Failed to get Operator %q", serviceMeshOperatorName)
 
-		restmapper := oc.AsAdmin().RESTMapper()
-		for _, ref := range operator.Status.Components.Refs {
-			mapping, err := restmapper.RESTMapping(ref.GroupVersionKind().GroupKind())
-			o.Expect(err).NotTo(o.HaveOccurred())
+			restmapper := oc.AsAdmin().RESTMapper()
+			for _, ref := range operator.Status.Components.Refs {
+				mapping, err := restmapper.RESTMapping(ref.GroupVersionKind().GroupKind())
+				o.Expect(err).NotTo(o.HaveOccurred())
 
-			err = oc.KubeFramework().DynamicClient.Resource(mapping.Resource).Namespace(ref.Namespace).Delete(context.Background(), ref.Name, metav1.DeleteOptions{})
-			o.Expect(err).Should(o.Or(o.Not(o.HaveOccurred()), o.MatchError(apierrors.IsNotFound, "IsNotFound")), "Failed to delete %s %q: %v", ref.GroupVersionKind().Kind, ref.Name, err)
+				err = oc.KubeFramework().DynamicClient.Resource(mapping.Resource).Namespace(ref.Namespace).Delete(context.Background(), ref.Name, metav1.DeleteOptions{})
+				o.Expect(err).Should(o.Or(o.Not(o.HaveOccurred()), o.MatchError(apierrors.IsNotFound, "IsNotFound")), "Failed to delete %s %q: %v", ref.GroupVersionKind().Kind, ref.Name, err)
+			}
 		}
 	})
 
 	g.It("Ensure OSSM and OLM related resources are created after creating GatewayClass", func() {
+		defer markTestDone(oc, ossmAndOLMResourcesCreated)
+
 		//check the catalogSource
 		g.By("Check OLM catalogSource, subscription, CSV and Pod")
 		waitCatalogErr := wait.PollUntilContextTimeout(context.Background(), 1*time.Second, 20*time.Minute, false, func(context context.Context) (bool, error) {
@@ -213,6 +240,7 @@ var _ = g.Describe("[sig-network-edge][OCPFeatureGate:GatewayAPIController][Feat
 	})
 
 	g.It("Ensure default gatewayclass is accepted", func() {
+		defer markTestDone(oc, defaultGatewayclassAccepted)
 
 		g.By("Check if default GatewayClass is accepted after OLM resources are successful")
 		errCheck := checkGatewayClass(oc, gatewayClassName)
@@ -220,6 +248,8 @@ var _ = g.Describe("[sig-network-edge][OCPFeatureGate:GatewayAPIController][Feat
 	})
 
 	g.It("Ensure custom gatewayclass can be accepted", func() {
+		defer markTestDone(oc, customGatewayclassAccepted)
+
 		customGatewayClassName := "custom-gatewayclass"
 
 		g.By("Create Custom GatewayClass")
@@ -247,6 +277,8 @@ var _ = g.Describe("[sig-network-edge][OCPFeatureGate:GatewayAPIController][Feat
 	})
 
 	g.It("Ensure LB, service, and dnsRecord are created for a Gateway object", func() {
+		defer markTestDone(oc, lbAndServiceAndDnsrecordAreCreated)
+
 		g.By("Ensure default GatewayClass is accepted")
 		errCheck := checkGatewayClass(oc, gatewayClassName)
 		o.Expect(errCheck).NotTo(o.HaveOccurred(), "GatewayClass %q was not installed and accepted", gatewayClassName)
@@ -270,6 +302,8 @@ var _ = g.Describe("[sig-network-edge][OCPFeatureGate:GatewayAPIController][Feat
 	})
 
 	g.It("Ensure HTTPRoute object is created", func() {
+		defer markTestDone(oc, httprouteObjectCreated)
+
 		g.By("Ensure default GatewayClass is accepted")
 		errCheck := checkGatewayClass(oc, gatewayClassName)
 		o.Expect(errCheck).NotTo(o.HaveOccurred(), "GatewayClass %q was not installed and accepted", gatewayClassName)
@@ -300,6 +334,8 @@ var _ = g.Describe("[sig-network-edge][OCPFeatureGate:GatewayAPIController][Feat
 	})
 
 	g.It("Ensure GIE is enabled after creating an inferencePool CRD", func() {
+		defer markTestDone(oc, gieEnabled)
+
 		errCheck := checkGatewayClass(oc, gatewayClassName)
 		o.Expect(errCheck).NotTo(o.HaveOccurred(), "GatewayClass %q was not installed and accepted", gatewayClassName)
 
@@ -775,4 +811,42 @@ func isOKD(oc *exutil.CLI) (bool, error) {
 		return true, nil
 	}
 	return false, nil
+}
+
+// annotationKeyForTest returns the key for an annotation on the default
+// gatewayclass that indicates whether the specified test is done.
+func annotationKeyForTest(testName string) string {
+	return fmt.Sprintf("test-%s-done", testName)
+}
+
+// markTestDone adds an annotation to the default gatewayclass that all the
+// GatewayAPIController tests use to indicate that a particular test has ended.
+// These annotations are used to determine whether it is safe to clean up the
+// gatewayclass and other shared resources.
+func markTestDone(oc *exutil.CLI, testName string) {
+	gwc, err := oc.AdminGatewayApiClient().GatewayV1().GatewayClasses().Get(context.Background(), gatewayClassName, metav1.GetOptions{})
+	o.Expect(err).NotTo(o.HaveOccurred())
+
+	if gwc.Annotations == nil {
+		gwc.Annotations = map[string]string{}
+	}
+	gwc.Annotations[annotationKeyForTest(testName)] = ""
+	_, err = oc.AdminGatewayApiClient().GatewayV1().GatewayClasses().Update(context.Background(), gwc, metav1.UpdateOptions{})
+	o.Expect(err).NotTo(o.HaveOccurred())
+}
+
+// checkAllTestsDone checks the annotations on the default gatewayclass that all
+// the GatewayAPIController tests use to determine whether all the tests are
+// done.
+func checkAllTestsDone(oc *exutil.CLI) bool {
+	gwc, err := oc.AdminGatewayApiClient().GatewayV1().GatewayClasses().Get(context.Background(), gatewayClassName, metav1.GetOptions{})
+	o.Expect(err).NotTo(o.HaveOccurred())
+
+	for _, testName := range testNames {
+		if _, ok := gwc.Annotations[annotationKeyForTest(testName)]; !ok {
+			return false
+		}
+	}
+
+	return true
 }
